+++ src/main/java/de/maxhenkel/better_respawn/IBetterDeathScreen.java
@@ -1,0 +1,7 @@
+package de.maxhenkel.better_respawn;
+
+public interface IBetterDeathScreen {
+
+    void setDelay(int delay);
+
+}